refactor,dataset: broke module into files and changed tdc implementation to remove numba
