+++ mokapot/model.py
@@ -29,6 +29,7 @@ from sklearn.svm import LinearSVC
 from typeguard import typechecked
 
 from mokapot import LinearPsmDataset
+from mokapot.dataset.base import update_labels
 
 LOGGER = logging.getLogger(__name__)
 
@@ -317,7 +318,9 @@ class Model:
             scores = scores[original_idx]
 
             # Update target
-            target = dataset._update_labels(scores, eval_fdr=self.train_fdr)
+            target = update_labels(
+                scores, dataset.target_values, eval_fdr=self.train_fdr
+            )
             target = target[shuffled_idx]
             num_passed.append((target == 1).sum())
 
@@ -329,10 +332,10 @@ class Model:
                 raise RuntimeError("Model performs worse after training.")
 
         # If the model performs worse than what was initialized:
-        if (
-            num_passed[-1] < (start_labels == 1).sum()
-            or num_passed[-1] < self.feat_pass
-        ):
+        best_feat_better = num_passed[-1] < self.feat_pass
+        start_better = num_passed[-1] < (start_labels == 1).sum()
+
+        if best_feat_better or start_better:
             if self.override:
                 LOGGER.warning("Model performs worse after training.")
             else:
@@ -598,8 +601,18 @@ def _get_starting_labels(dataset: LinearPsmDataset, model):
 
     else:
         feat = dataset.features[model.direction].values
-        desc_labels = dataset._update_labels(feat, model.train_fdr, desc=True)
-        asc_labels = dataset._update_labels(feat, model.train_fdr, desc=False)
+        desc_labels = update_labels(
+            feat,
+            dataset.target_values,
+            model.train_fdr,
+            desc=True,
+        )
+        asc_labels = update_labels(
+            feat,
+            dataset.target_values,
+            model.train_fdr,
+            desc=False,
+        )
         best_feat = feat
 
         desc_pass = (desc_labels == 1).sum()
