+++ mokapot/picked_protein.py
@@ -4,8 +4,9 @@ confidence estimates.
 """
 
 import logging
-import pandas as pd
+
 import numpy as np
+import pandas as pd
 
 from mokapot import utils
 from mokapot.peptides import match_decoy
