+++ mokapot/tabular_data/__init__.py
@@ -1,21 +1,21 @@
 from .base import (
-    auto_finalize,
-    remove_columns,
     BufferType,
     ColumnMappedReader,
     ColumnSelectReader,
     DataFrameReader,
     TabularDataReader,
     TabularDataWriter,
+    auto_finalize,
+    remove_columns,
 )
 from .csv import CSVFileReader, CSVFileWriter
-from .parquet import ParquetFileWriter, ParquetFileReader
+from .parquet import ParquetFileReader, ParquetFileWriter
+from .sqlite import ConfidenceSqliteWriter, SqliteWriter
 from .streaming import (
-    join_readers,
-    merge_readers,
     BufferedWriter,
     ComputedTabularDataReader,
     JoinedTabularDataReader,
     MergedTabularDataReader,
+    join_readers,
+    merge_readers,
 )
-from .sqlite import ConfidenceSqliteWriter, SqliteWriter
