+++ mokapot/tabular_data/target_decoy_writer.py
@@ -1,5 +1,5 @@
-from typing import Sequence
 import warnings
+from typing import Sequence
 
 import numpy as np
 import pandas as pd
