+++ mokapot/tabular_data/traditional_pin.py
@@ -1,5 +1,5 @@
-from pathlib import Path
 from io import StringIO
+from pathlib import Path
 
 import pandas as pd
