+++ pyproject.toml
@@ -6,9 +6,7 @@ requires = ["setuptools>=61", "wheel", "setuptools_scm[toml]>=3.4"]
 dev = [
   "twine>=6.0.1",
   "pre-commit>=2.7.1",
-  "ruff>=0.8.1",
-  "pytest>=8.2.2",
-  "pytest-cov>=6.0.0"
+  "ruff>=0.8.1"
 ]
 docs = [
   "numpydoc>=1.0.0",
@@ -17,6 +15,14 @@ docs = [
   "nbsphinx>=0.7.1",
   "ipykernel>=5.3.0",
   "recommonmark>=0.5.0"
+]
+profile = [
+  "scalene",
+  "memray"
+]
+test = [
+  "pytest>=8.2.2",
+  "pytest-cov>=6.0.0"
 ]
 
 [project]
@@ -31,7 +37,6 @@ dependencies = [
   "numpy>=1.4.0,<2.0.0",  # required for triqler https://github.com/statisticalbiotechnology/triqler/pull/34
   "pandas>=1.0.3",
   "scikit-learn>=0.22.1",
-  "numba>=0.48.0",
   "lxml>=4.6.2",
   "triqler>=0.6.2",
   "joblib>=1.1.0",
@@ -75,7 +80,7 @@ docstring-code-format = true
 preview = true
 
 [tool.ruff.lint]
-select = ["E", "F", "T20"]  # T20 is for print() statements.
+select = ["E", "F", "T20", "I"]
 
 [tool.ruff.lint.per-file-ignores]
 "__init__.py" = ["F401"]
