+++ tests/unit_tests/test_brew.py
@@ -7,7 +7,7 @@ import pytest
 from sklearn.tree import DecisionTreeClassifier
 
 import mokapot
-from mokapot import PercolatorModel, Model
+from mokapot import LinearPsmDataset, Model, PercolatorModel
 
 np.random.seed(42)
 
@@ -23,6 +23,49 @@ def test_brew_simple(psms_ondisk, svm):
     models, scores = mokapot.brew([psms_ondisk], svm, test_fdr=0.05)
     assert len(models) == 3
     assert isinstance(models[0], PercolatorModel)
+
+
+def test_brew_positive_negative(psm_df_builder, svm):
+    """A small-ish PSM dataset"""
+    data = psm_df_builder(1000, 1000, score_diffs=[5.0])
+
+    psms_positive = LinearPsmDataset(
+        psms=data.df,
+        target_column="target",
+        spectrum_columns=["specid"],
+        peptide_column="peptide",
+        feature_columns=list(data.score_cols),
+        copy_data=True,
+        rng=42,
+    )
+    data_negative = data.df.copy()
+    for c in data.score_cols:
+        data_negative[c] *= -1
+
+    psms_negative = LinearPsmDataset(
+        psms=data_negative,
+        target_column="target",
+        spectrum_columns=["specid"],
+        peptide_column="peptide",
+        feature_columns=list(data.score_cols),
+        copy_data=True,
+        rng=42,
+    )
+
+    pos_models, pos_scores = mokapot.brew(
+        [psms_positive],
+        PercolatorModel(train_fdr=0.05, max_iter=10, rng=2),
+        test_fdr=0.05,
+        rng=2,
+    )
+    neg_models, neg_scores = mokapot.brew(
+        [psms_negative],
+        PercolatorModel(train_fdr=0.05, max_iter=10, rng=2),
+        test_fdr=0.05,
+        rng=2,
+    )
+
+    assert np.allclose(pos_scores[0], neg_scores[0])
 
 
 def test_brew_simple_parquet(psms_ondisk_from_parquet, svm):
@@ -78,6 +121,12 @@ def test_brew_folds(psms_ondisk, svm):
 @pytest.mark.slow
 def test_brew_seed(psms_ondisk, svm):
     """Test that (not) changing the split selection seed works"""
+    # Set logger level to debug
+
+    import logging
+
+    logging.getLogger("mokapot").setLevel(logging.DEBUG)
+
     folds = 3
     seed = 0
     psms_ondisk_b = copy.copy(psms_ondisk)
@@ -99,6 +148,11 @@ def test_brew_seed(psms_ondisk, svm):
     models_c, scores_c = mokapot.brew(
         [psms_ondisk_c], svm, test_fdr=0.05, folds=folds, rng=seed + 2
     )
+
+    # Scores can be zero if the seed is the same BUT the model cannot
+    # tell between targets and decoys. This should not happen in this test
+    # and would usually mean there is a bug elsewhere.
+    assert scores_a[0].max() > 0, "Scores are all zero"
     assert len(models_c) == folds
     assert not (np.array_equal(scores_a[0], scores_c[0])), (
         "Results were identical with different seed!"
@@ -141,6 +195,8 @@ def test_brew_seed_parquet(psms_ondisk_from_parquet, svm):
         rng=seed + 2,
     )
     assert len(models_c) == folds
+    # This happens when the internal seeding mechanism is not
+    # correctly setting the seed internally.
     assert not (np.array_equal(scores_a[0], scores_c[0])), (
         "Results were identical with different seed!"
     )
@@ -148,6 +204,7 @@ def test_brew_seed_parquet(psms_ondisk_from_parquet, svm):
 
 def test_brew_test_fdr_error(psms_ondisk, svm):
     """Test that we get a sensible error message"""
+    # Q: When what happens???
     with pytest.raises(RuntimeError) as err:
         mokapot.brew([psms_ondisk], svm, test_fdr=0.001, rng=2)
     assert "Failed to calibrate" in str(err)
