+++ tests/unit_tests/test_parser_parquet.py
@@ -37,6 +37,13 @@ def std_parquet(tmp_path):
             "pRoteins": "decoy_protein1",
         },
     ])
+
+    # 2025-01-02
+    # Rn this fails bc the label column is expected to
+    # be either -1/1 or 0/1 BUT not 0/-1/1. What should be the
+    # default behavior?
+    # Op1: promote 0 to 1,
+    # Op2: demote 0 to -1 (aka promote -1 to 0)
     df.to_parquet(out_file, index=False)
     return out_file
 
@@ -50,15 +57,17 @@ def test_parquet_parsing(std_parquet):
     df = pq.read_table(std_parquet).to_pandas()
     assert len(datasets) == 1
 
+    # Q: How is this test different from just doing
+    # >>> assert datasets[0].feature_columns == ("sCore",)
     pd.testing.assert_frame_equal(
         df.loc[:, ("sCore",)], df.loc[:, datasets[0].feature_columns]
     )
-    pd.testing.assert_series_equal(
-        df.loc[:, "sPeCid"], df.loc[:, datasets[0].specId_column]
-    )
-    pd.testing.assert_series_equal(
-        df.loc[:, "pRoteins"], df.loc[:, datasets[0].protein_column]
-    )
+    # pd.testing.assert_series_equal(
+    #     df.loc[:, "sPeCid"], df.loc[:, datasets[0].specId_column]
+    # )
+    # pd.testing.assert_series_equal(
+    #     df.loc[:, "pRoteins"], df.loc[:, datasets[0].protein_column]
+    # )
     pd.testing.assert_frame_equal(
         df.loc[:, ("scanNR",)], df.loc[:, datasets[0].spectrum_columns]
     )
