+++ tests/unit_tests/test_proteins.py
@@ -1,8 +1,9 @@
 """Test that we can parse a FASTA file correctly"""
 
 import pytest
+
 import mokapot
-from mokapot import read_fasta, digest, make_decoys
+from mokapot import digest, make_decoys, read_fasta
 
 
 @pytest.fixture
